Add switch between Luring / Daring functionality
