+++ pasta/siakad_auto_attendance/utils.py
@@ -67,6 +67,12 @@ def add_entry_to_config(name: str, value: str):
 
     save_to_config(config)
 
+def yn_choice(message, default='y'):
+    choices = 'Y/n' if default.lower() in ('y', 'yes') else 'y/N'
+    choice = input("%s (%s) " % (message, choices))
+    values = ('y', 'yes', '') if choices == 'Y/n' else ('y', 'yes')
+    return choice.strip().lower() in values
+
 ###Connection checker is still under construction###
 async def checkConnection():
     while True:
